refactor: change interfaces place in promocode
