--- src/components/PromoCode/displayPromoCodes/PromoCodeInterfaces/PromoCodeInterfaces.ts
@@ -1,4 +1,0 @@
-export interface PromoCodeProperties {
-  code: string;
-  description: string;
-}
