+++ src/components/PromoCode/displayPromoCodes/PromoCodes/PromoCodes.ts
@@ -1,4 +1,9 @@
-const promoCodes = [
+export interface PromoCodeProperties {
+  code: string;
+  description: string;
+}
+
+const promoCodes: PromoCodeProperties[] = [
   { code: 'POSTER15', description: 'All-15%' },
   {
     code: 'POSTER20',
